refactor(windows): move CLI things inside client.rs
